not saving training output during hyperparameter optimization now

+++ train/general_training.py
@@ -12,13 +12,14 @@ from train.trainer import AdversarialAutoencoderTrainer, AutoencoderTrainer, Dif
 from utils.config_parser import ConfigParser
 
 
-def train(config: Dict[str, Any]):
+def train(config: Dict[str, Any], save_output: bool = True):
     train_config = config["training"]
     out_path = Path(train_config["out_path"]).expanduser()
 
-    if out_path.exists():
-        shutil.rmtree(out_path, ignore_errors=True)
-    out_path.mkdir()
+    if save_output:
+        if out_path.exists():
+            shutil.rmtree(out_path, ignore_errors=True)
+        out_path.mkdir()
 
     device = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")
 
@@ -65,7 +66,7 @@ def train(config: Dict[str, Any]):
             val_dataset=dataset_val
         )
     else:
-        model = model_class(config["model"]["params"], train_config, device=device)
+        model = model_class(config["model"]["params"], train_config, device=device, save_output=save_output)
 
     best_val_loss = np.inf
     epochs_without_improvement = 0
@@ -89,7 +90,9 @@ def train(config: Dict[str, Any]):
         if model.metrics_logger.get_last("val_loss") <= best_val_loss:
             best_val_loss = model.metrics_logger.get_last("val_loss")
             epochs_without_improvement = 0
-            torch.save(model.get_model().state_dict(), out_path.joinpath("best_model.pt"))
+
+            if save_output:
+                torch.save(model.get_model().state_dict(), out_path.joinpath("best_model.pt"))
             best_state_dict = model.get_model().state_dict().copy()
         else:
             epochs_without_improvement += 1
@@ -97,7 +100,9 @@ def train(config: Dict[str, Any]):
             if epochs_without_improvement > train_config["break_criterion"]:
                 break
 
-    torch.save(model.get_model().state_dict(), out_path.joinpath("final_model.pt"))
+    if save_output:
+        torch.save(model.get_model().state_dict(), out_path.joinpath("final_model.pt"))
+        model.metrics_logger.save()
 
     return best_state_dict, best_val_loss
